#   Remove strong reference of XNamespace from XName