Fixes https://github.com/dotnet/corefx/issues/36404
XNamespace has a hashtable of XNames keeping all Names bound to it, also XNames has strong reference back to its XNamespace.When creating lots of XNames within same XNamespace its XNames Hastable grows endlessly until fill out memory. Even XNamespaces are WeakReferences it stays in memory forever because they have strong references to its XNames. With this change removing that strong reference from XName