@buyaa-n 2019-11-26
> Nit, up to you, but consider "git rebase -i" such that you end up with 1x commit to do the cleanup (eg., using WeakReference<T>) and 1x commit to do the fix. Then it's easy to review them separately.

It cannot really cleanly divided like that, all mixed. Anyways I might need to create new PR due to recreated fork, will keep that in mind if that happens