@buyaa-n 2019-12-05
I had some suggestion from Adam about how to have more stable env when running perf tests. Had more stable test results but the result was not good as I was hoping:
summary:
better: 1, geomean: 1.054
worse: 9, geomean: 1.128
total diff: 10

| Slower                                                    | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Xml.Linq.Perf_XElement.GetElement                  |      1.22 |            48.98 |            59.85 |         |
| System.Xml.Linq.Perf_XName.CreateElement                  |      1.16 |            33.98 |            39.28 |         |
| System.Xml.Linq.Perf_XElement.GetElementWithNamespace     |      1.15 |            94.22 |           108.24 |         |
| System.Xml.Linq.Perf_XElement.GetAttribute                |      1.13 |            35.03 |            39.66 |         |
| System.Xml.Linq.Perf_XElement.CreateWithElements          |      1.13 |           107.99 |           121.92 |         |
| System.Xml.Linq.Perf_XElement.GetValue                    |      1.12 |            72.12 |            80.58 |         |
| System.Xml.Linq.Perf_XElement.CreateElement               |      1.12 |            57.47 |            64.11 |         |
| System.Xml.Linq.Perf_XElement.CreateElementsWithNamespace |      1.08 |           159.08 |           171.84 | bimodal |
| System.Xml.Linq.Perf_XElement.CreateElementWithNamespace  |      1.06 |           105.82 |           112.01 |         |

| Faster                                               | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Xml.Linq.Perf_XName.NonEmptyNameSpaceToString |      1.05 |            37.19 |            35.29 |         |

Now I cannot say it is acceptable. 

Anyways gonna close this PR as its not accessable anymore. Will create another PR if we want to fix the issue with this performance degression.